network base connfigutation by installer
basic network configurations will be executed by installer (user defined)

+++ install/common.sh
@@ -23,12 +23,104 @@
 weneco_dir="/etc/weneco"
 webroot_dir="/var/www/html/weneco"
 weneco_user="www-data"
-php_package="php7.0-cgi" 
-#php_package="php5-cgi"
+
+
+# TEXT COLORS
+rd="\033[1;31m"
+gn="\033[1;32m"
+ye="\033[1;33m"
+nc="\033[0m"
+
+#------------------------------------
+#    CONFIG INSTALL MENU
+#------------------------------------
+
+# CONFIGURE INSTALLATION BY USER
+function config_install(){
+    echo -e ""
+    echo -e "${gn}------------------------------ ${nc}"
+    echo -e "${gn}  Installation configuration ${nc}"
+    echo -e "${gn}------------------------------${nc}"
+    echo -e "${ye}Services wicd, dhcpcd, conman and NetworkManager will be DISABLED${nc}"
+    echo -e "${ye}It's recommed to do installation local not over ssh or network ${nc}"
+    echo -e "Old configuration files will be backed up to install directory"
+    echo -e ""
+    # PHP-Version
+    echo -e "Which PHP-Version should be installed?"
+    echo -e "1) PHP 5"
+    echo -e "2) PHP 7 (recommed)"
+    echo -n "Choose Version [1/2] :"
+    read answer
+    if [[ $answer == "1" ]]; then
+        php_package="php5-cgi"
+    else
+        php_package="php7.0-cgi" 
+    fi
+    # weneco_dir
+    echo -n "Change install directory: '${weneco_dir}'? [y/N]:"
+    read answer
+    if [[ $answer == "y" ]]; then
+        echo -n "enter new directory: "
+        read new_dir
+        weneco_dir=$new_dir
+    fi
+    if [ -d "$weneco_dir" ]; then
+        echo -e "content of '${weneco_dir}' will be moved to ${weneco_dir}.date" 
+    fi
+    # webroot_dir
+    echo -n "Change Lighttpd directory: '${webroot_dir}'? [y/N]:"
+    read answer
+    if [[ $answer == "y" ]]; then
+        echo -n "enter new directory: "
+        read new_dir
+        webroot_dir=$new_dir
+    fi
+    if [ -d "$webroot_dir" ]; then
+        echo -e "content of '${webroot_dir}' will be moved to ${webroot_dir}.date" 
+    fi
+    
+    echo -e ""
+    echo -e "${ye}Summary:${nc}"
+    echo -e "-----------------"    
+    echo -e "PHP-Version: $php_package"
+    echo -e "install-directory: '$weneco_dir'"
+    echo -e "website-directory: '$webroot_dir'"
+    echo -e ""
+    echo -n "Complete installation with these values? [y/N]: "
+    read answer
+    if [[ $answer != "y" ]]; then
+        echo -e "${rd}Installation aborted by user${nc}"
+        exit 0
+    fi
+}
+
+#------------------------------------
+#          LOG FUNCTIONS
+#------------------------------------
 
 # Outputs a log line
 function log() {
-    echo -e "\033[1;32mWeNeCo Install: $*\033[m"
+    echo -e "${gn}WeNeCo Install: $${nc}"
+}
+
+# Outputs a log line without newline
+function log_ne(){
+    echo -ne "\033[1;30mWeNeCo Install:${nc} $*"
+}
+
+# append OK
+function log_ok(){
+    echo -e "${gn} OK${nc} $*"
+}
+
+# append FAILED
+function log_failed(){
+    echo -e "${rd} FAILED${nc} $*"
+}
+
+# append OK
+function log_warn(){
+    echo -e "${ye} WARNING${nc} $*"
 }
 
 # Outputs install error log line and exits with status code 1
@@ -36,6 +128,10 @@ function install_error() {
     echo -e "\033[1;37;41mWeNeCo Install Error: $*\033[m"
     exit 1
 }
+
+#------------------------------------
+#          SYSTEM FUNCTIONS
+#------------------------------------
 
 # CHECK SYSTEM 
 function check_system() {
@@ -52,28 +148,9 @@ function update_system(){
     sudo apt-get update || install_error "Unable to update package list"
 }
 
-# CONFIGURE INSTALLATION BY USER
-function config_install(){
-    log "CONFIGURE INSTALLATION"
-    echo -e "Install directory: ${weneco_dir}"
-    echo -e "Lighttpd directory: ${webroot_dir}"
-    echo -e "\033[1;33mServices wicd, dhcpcd, conman and NetworkManager will be DISABLED\033[m"
-    echo -e "\033[1;33mIt's recommed to do installation local not over network, ssh\033[m"
-    echo -e "Old configuration files will be backed up to install directory"
-    if [ -d "$webroot_dir" ]; then
-        echo -e "content of '${webroot_dir}' will be moved to ${webroot_dir}.date" 
-    fi
-    echo -n "Complete installation with these values? [y/N]: "
-    read answer
-    if [[ $answer != "y" ]]; then
-        echo -e "\033[1;31mInstallation aborted by user\033[m"
-        exit 0
-    fi
-}
-
 # CREATE DIRECTORIES
 function create_directories(){
-    install_log "Creating RaspAP directories"
+    log_ne "Creating WeNeCo directories"
     # Main directory
     if [ -d "$weneco_dir" ]; then
         mv $weneco_dir "$weneco_dir.`date +%F-%R`" || install_error "Unable to move old '$weneco_dir' out of the way"
@@ -85,11 +162,13 @@ function create_directories(){
     
     # Backup directory
     sudo mkdir -p "$weneco_dir/network"
+    
+    log_ok
 }
     
 # BACKUP OLD CONFIGURE
 function backup_config(){
-    log "backup old configuration files"
+    log_ne "backup old configuration files"
     if [ -f /etc/network/interfaces ]; then
         sudo cp /etc/network/interfaces "$weneco_dir/backups/interfaces.`date +%F-%R`"
     fi
@@ -109,6 +188,8 @@ function backup_config(){
     if [ -f /etc/rc.local ]; then
         sudo cp /etc/rc.local "$weneco_dir/backups/rc.local.`date +%F-%R`"
     fi
+    
+    log_ok
 }
 
 # INSTALL PACKAGE
@@ -116,10 +197,11 @@ function install_package(){
     state=$(eval "systemctl is-active $1")
     if [ $state == "inactive"  ]
     then
-        log "instaling $1"
+        log_ne "instaling $1"
         eval "sudo apt-get install $1" || install_error "Unable to install $1"
+        log_ok
     else
-        log "$1 already installed"
+        log_warn "$1 already installed - skipped"
     fi
 }
 
@@ -137,17 +219,19 @@ function download_lates(){
         sudo mv $webroot_dir "$webroot_dir.`date +%F-%R`" || install_error "Unable to remove old webroot directory"
     fi
 
-    install_log "Cloning latest files from github"
+    log_ne "Cloning latest files from github"
     git clone https://github.com/dawildde/WeNeCo /tmp/weneco || install_error "Unable to download files from github"
     sudo mv /tmp/weneco $webroot_dir || install_error "Unable to move webgui to web root"
+    log_ok
 }
 
 # MOVE FILES TO APP DIRECTORIES
 function move_files(){
     if [ -d "$webroot_dir" ] && [ -d "$weneco_dir" ]; then
-        log "moving files"
+        log_ne "moving files"
         echo "'$webroot_dir && '$weneco_dir'"
         sudo mv "$webroot_dir/config" "$weneco_dir" || install_error "Unable to move files to $weneco_dir"
+        log_ok
     else
         install_error "WeNeCo directories are not existing. Please reinstall"
     fi
@@ -155,8 +239,22 @@ function move_files(){
 
 # UPDATE SYSTEM FILES
 function overwrite_systemfiles(){
-    sudo cp "$weneco_dir/config/interfaces" /etc/network/interfaces
-    sudo rm /etc/resolv.conf
+    log_ne "overwrite system-files"
+    sudo cp "$weneco_dir/config/interfaces" /etc/network/interfaces  || install_error "Unable to overwrite '/etc/network/interfaces/'"
+    if [ -f /etc/resolv.conf ]; then
+        sudo rm /etc/resolv.conf
+    fi
+    log_ok
+}
+
+# COPY NETWORK SETTINGS
+function overwrite_networkfiles(){
+    if [ -f "$weneco_dir/config/device0.network" ]; then
+        log_ne "copy network files to '/etc/systemd/network/'"
+        sudo mv "/etc/systemd/network/device*.network" "$weneco_dir/backups/network.`date +%F-%R`/" 2>/dev/nul
+        sudo cp $weneco_dir/config/device*.network /etc/systemd/network/ || install_error "Unable to overwrite network files in '/etc/systemd/network/'"
+        log_ok
+    fi
 }
 
 # DISABLE SERVICE
@@ -202,6 +300,7 @@ function install_weneco(){
     download_latest
     move_files
     patch_all
+    configure_network
     overwrite_systemfiles
     disable_services 
     enable_systemd
@@ -209,6 +308,10 @@ function install_weneco(){
 
 # UPDATE
 function update_weneco(){
+    echo -e "${gn}----------------------- ${nc}"
+    echo -e "${gn}    Updating WeNeCo ${nc}"
+    echo -e "${gn}-----------------------${nc}"
     download_latest
+    move_files
 }
